+++ src/Settings.cs
@@ -3,7 +3,7 @@ using MCM.Abstractions.Attributes;
 using MCM.Abstractions.Attributes.v2;
 using MCM.Abstractions.Settings.Base.Global;
 
-namespace Pacemaker
+namespace CampaignPacer
 {
 	public class Settings : AttributeGlobalSettings<Settings>
 	{