+++ src/CustomSaveableTypeDefiner.cs
@@ -10,17 +10,6 @@ namespace Pacemaker
 
         public CustomSaveableTypeDefiner() : base(SaveBaseId) { }
 
-        protected override void DefineClassTypes()
-        {
-            AddClassDefinition(typeof(SimpleTime), 1);
-            AddClassDefinition(typeof(SavedValues), 2);
-        }
-
-        //protected override void DefineContainerDefinitions()
-        //{
-        //    ConstructContainerDefinition(typeof(Dictionary<string, long>));
-        //    ConstructContainerDefinition(typeof(Dictionary<string, double>));
-        //    ConstructContainerDefinition(typeof(Dictionary<string, string>));
-        //}
+        protected override void DefineClassTypes() => AddClassDefinition(typeof(SavedValues), 1);
     }
 }